initial commit for adding /flight endpoint
