+++ package.json
@@ -34,21 +34,21 @@
     "@types/dotenv": "^6.1.1",
     "@types/dotenv-flow": "^3.2.0",
     "@types/express": "^4.17.13",
-    "@types/jest": "^27.0.1",
+    "@types/jest": "^27.0.2",
     "@types/luxon": "^2.0.4",
-    "@types/node": "^16.9.1",
+    "@types/node": "^16.9.4",
     "@types/seedrandom": "^3.0.1",
     "@types/supertest": "^2.0.11",
-    "@typescript-eslint/eslint-plugin": "^4.31.1",
-    "@typescript-eslint/parser": "^4.31.1",
-    "babel-jest": "^27.1.1",
+    "@typescript-eslint/eslint-plugin": "^4.31.2",
+    "@typescript-eslint/parser": "^4.31.2",
+    "babel-jest": "^27.2.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-airbnb-typescript": "^14.0.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jest": "^24.4.0",
-    "jest": "^27.2.0",
+    "eslint-plugin-jest": "^24.4.2",
+    "jest": "^27.2.1",
     "prettier": "^2.4.0",
     "rimraf": "^3.0.2",
     "supertest": "^6.1.6",
